requirements.txt
streamlit>=1.26.0
pinecone==6.0.2
python-dotenv>=1.0.0
protobuf>=4.24.0
numpy>=1.20.0 
openai